jpsc_sap_api/requirements.txt
fastapi
pyodbc
sqlalchemy
uvicorn
FastAPI-SQLAlchemy
python-dotenv